Move `Derivations` class to its own file
